requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
openai
anthropic
alembic
pytest
httpx
pytest-cov
ruff
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
bcrypt<4
slowapi
redis
email-validator
